Move node.py to __init__.py
